+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - use celery as queue to run background tasks ([#72])
 - implement web map service client ([#95])
 - add resources (compass and globes as SVG graphics) for map generation ([#101])
+- add module containing the paper format specs ([#99])
 
 ### Bug Fixes
 
@@ -37,4 +38,5 @@
 [#72]: https://github.com/GIScience/sketch-map-tool/pull/72
 [#73]: https://github.com/GIScience/sketch-map-tool/pull/73
 [#95]: https://github.com/GIScience/sketch-map-tool/pull/95
+[#99]: https://github.com/GIScience/sketch-map-tool/pull/99
 [#101]: https://github.com/GIScience/sketch-map-tool/pull/101